refactor: improve FileViewer types organization and UploadFolderManager return consistency
